+++ apps/backend/src/modules/courses/course.repository.ts
@@ -6,6 +6,7 @@ import { SortOrder } from "~/libs/enums/enums.js";
 import { DatabaseTableName } from "~/libs/modules/database/database.js";
 import { HTTPCode } from "~/libs/modules/http/http.js";
 import {
+	type PaginationRequestDto,
 	type PaginationResponseDto,
 	type Repository,
 } from "~/libs/types/types.js";
@@ -161,11 +162,15 @@ class CourseRepository implements Repository<CourseEntity> {
 			: null;
 	}
 
-	public async findAll(): Promise<{ items: CourseEntity[] }> {
-		const courses = await this.courseModel
+	public async findAll({
+		count,
+		page,
+	}: PaginationRequestDto): Promise<PaginationResponseDto<CourseEntity>> {
+		const { results: courses, total } = await this.courseModel
 			.query()
 			.withGraphFetched("vendor")
 			.orderBy("id", SortOrder.ASC)
+			.page(page, count)
 			.execute();
 
 		return {
@@ -190,6 +195,7 @@ class CourseRepository implements Repository<CourseEntity> {
 					vendorId: course.vendorId,
 				});
 			}),
+			total,
 		};
 	}