+++ apps/backend/src/modules/courses/course.service.ts
@@ -1,5 +1,10 @@
 import { HTTPCode } from "~/libs/enums/enums.js";
+import { convertPageToZeroIndexed } from "~/libs/helpers/helpers.js";
 import { type OpenAI } from "~/libs/modules/open-ai/open-ai.js";
+import {
+	type PaginationRequestDto,
+	type PaginationResponseDto,
+} from "~/libs/types/types.js";
 import {
 	CourseSectionEntity,
 	type CourseSectionRepository,
@@ -220,15 +225,22 @@ class CourseService {
 		return entity.toObject();
 	}
 
-	public async findAll(): Promise<CourseSearchGetAllResponseDto> {
-		const { items } = await this.courseRepository.findAll();
+	public async findAll({
+		count,
+		page,
+	}: PaginationRequestDto): Promise<PaginationResponseDto<CourseDto>> {
+		const { items: courses, total } = await this.courseRepository.findAll({
+			count,
+			page: convertPageToZeroIndexed(page),
+		});
 
-		const courses = items.map((entity) => ({
-			...entity.toObject(),
-			hasUserCourse: false,
-		}));
-
-		return { courses };
+		return {
+			items: courses.map((entity) => ({
+				...entity.toObject(),
+				hasUserCourse: false,
+			})),
+			total,
+		};
 	}
 
 	public async findAllByVendor(