+++ apps/frontend/src/modules/courses/slices/courses.slice.ts
@@ -33,6 +33,7 @@ type State = {
 	recommendedDataStatus: ValueOf<typeof DataStatus>;
 	searchDataStatus: ValueOf<typeof DataStatus>;
 	searchedCourses: CourseSearchResponseDto[];
+	totalCoursesCount: number;
 };
 
 const initialState: State = {
@@ -45,6 +46,7 @@ const initialState: State = {
 	recommendedDataStatus: DataStatus.IDLE,
 	searchDataStatus: DataStatus.IDLE,
 	searchedCourses: [],
+	totalCoursesCount: 0,
 };
 
 const { actions, name, reducer } = createSlice({
@@ -110,10 +112,14 @@ const { actions, name, reducer } = createSlice({
 			state.addedVendorCourseDataStatuses[vendorCourseId] =
 				DataStatus.FULFILLED;
 		});
-		builder.addCase(getAll.fulfilled, (state, action) => {
-			state.allCourses = action.payload;
-			state.allCoursesDataStatus = DataStatus.FULFILLED;
-		});
+		builder.addCase(
+			getAll.fulfilled,
+			(state, { payload: { items, total } }) => {
+				state.allCourses = items;
+				state.totalCoursesCount = total;
+				state.allCoursesDataStatus = DataStatus.FULFILLED;
+			},
+		);
 		builder.addCase(getAll.pending, (state) => {
 			state.allCoursesDataStatus = DataStatus.PENDING;
 		});
@@ -123,7 +129,15 @@ const { actions, name, reducer } = createSlice({
 
 		builder.addCase(
 			deleteById.fulfilled,
-			(state, { meta: { arg: courseId }, payload }) => {
+			(
+				state,
+				{
+					meta: {
+						arg: { courseId },
+					},
+					payload,
+				},
+			) => {
 				if (payload) {
 					state.allCourses = state.allCourses.filter((course) => {
 						return course.id !== courseId;
@@ -136,7 +150,14 @@ const { actions, name, reducer } = createSlice({
 		);
 		builder.addCase(
 			deleteById.pending,
-			(state, { meta: { arg: courseId } }) => {
+			(
+				state,
+				{
+					meta: {
+						arg: { courseId },
+					},
+				},
+			) => {
 				state.courseToDataStatus[courseId] = {
 					deleteDataStatus: DataStatus.PENDING,
 				};
@@ -144,7 +165,14 @@ const { actions, name, reducer } = createSlice({
 		);
 		builder.addCase(
 			deleteById.rejected,
-			(state, { meta: { arg: courseId } }) => {
+			(
+				state,
+				{
+					meta: {
+						arg: { courseId },
+					},
+				},
+			) => {
 				state.courseToDataStatus[courseId] = {
 					deleteDataStatus: DataStatus.REJECTED,
 				};