+++ apps/frontend/src/modules/users/libs/types/types.ts
@@ -1,6 +1,5 @@
 export {
 	type UserAuthResponseDto,
-	type UserGetAllResponseDto,
 	type UserProfileRequestDto,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,