+++ packages/shared/src/modules/users/libs/validation-schemas/validation-schemas.ts
@@ -1,3 +1,4 @@
+export { userGetAllQuery } from "./user-get-all-query.validation-schema.js";
 export { userIdParameters } from "./user-id-parameters.validation-schema.js";
 export { userProfile } from "./user-profile.validation-schema.js";
 export { userSignIn } from "./user-sign-in.validation-schema.js";